Create Cybersecurity
Here is an example code for implementing OAuth authentication using the Flask web framework in Python:

This code demonstrates the basic flow of an OAuth authentication. When the user visits the /login endpoint, they are redirected to the Facebook authorization page. After the user grants permission, they are redirected back to the /authorized endpoint, where the code exchanges the authorization code for an access token. The access token can then be used to make API calls to retrieve the user's profile information.

Note that this is just a simple example, and you may need to modify it to meet the specific requirements of your application. Additionally, you may need to install the flask-oauthlib library by running pip install flask-oauthlib.